pyproject.toml
[project]
name = "django-getpaid-elavon"
version = "0.1.0"
description = "django-getpaid processor for Elavon"
authors = [
    { name = "Order Group", email = "adam.przyborowski@ordergroup.pl" }
]
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = { text = "MIT" }
repository = "https://github.com/ordergroup/og-django-getpaid-elavon"
dependencies = [
    "django-getpaid>=2.3.0,<3.0",
    "Django>=5.2",
    "swapper>=1.3.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
test = [
    "pytest>=6.2.4",
    "codecov>=2.1.11",
    "coverage>=5.5",
    "requests-mock>=1.9.3",
    "pytest-cov>=2.12.1",
    "pytest-django>=4.4.0",
    "pytest-factoryboy>=2.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=2.16.0",
    "factory-boy>=3.2.1",
    "ruff>=0.8.0",
    "tox>=3.24.4",
    "ipdb>=0.13.9",
]
test = [
    "pytest>=6.2.4",
    "codecov>=2.1.11",
    "coverage>=5.5",
    "requests-mock>=1.9.3",
    "pytest-cov>=2.12.1",
    "pytest-django>=4.4.0",
]

[tool.uv]
package = true
override-dependencies = [
    "pendulum>=3.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["getpaid_elavon"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["getpaid_elavon"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
